Some work in progress on improving MonthlyBuild.
Basically, tagging the abstracts seizes a ton of memory and doesn't release it, meaning the remaining tasks thrash so much they take hours or days to complete instead of minutes.  I think that's because `@AllAbstracts` holds onto a big list of all of the abstracts complete with MESH tags, and is never deallocated.  Unfortunately, it's hard to test right now to see whether I need to break the build into multiple parts or whether I can just dump the reference to the abstracts list.